Use templates for netmask/hostmask
